Added UID and GID support to OS X. Also updated psutil.py and _pslinux.py to read the UID and GID at the time of get_process_info() - since this information is part of the same kernel structure we use to get process name etc. then it makes sense to call it all at once instead of separately for performance.
Updated the ProcessInfo class to have uid and gid members and updated the 
Process class to retrieve this information from the ProcessInfo class. 
This also does away with the need to cache the information, since it will
exist for the life of the ProcessInfo object.

Lastly, updated the infoTuple returned from Windows and OS X so that it 
returns None for uid and gid if not available. Since this is not implemented
on Windows platform yet, Windows will always return None for UID/GID for now.