SwapTotal should be used for total virtual memory, not VmalloTotal.
See http://lkml.indiana.edu/hypermail/linux/kernel/0506.1/0992.html 

"The kernel has one gigabyte to map on x86. In your case 256M is used for
RAM and the most of the rest can go to vmalloc. VmallocTotal merely
indicates the amount of virtual memory that can be used for vmalloc'ed
(non-contigous) mappings."